+++ src/app.js
@@ -3,8 +3,8 @@
 require('dotenv').config();
 const express = require('express');
 
-const loginRouter = require('./login-router/login');
-const appLogicRouter = require('./app-logic-router/app-logic-router');
+const loginRouter = require('./routers/login-router/login');
+const appLogicRouter = require('./routers/app-logic-router/app-logic-router');
 
 // -----------------------------------------------------------------------
 const app = express();
@@ -20,12 +20,6 @@ app.use(require('express-session')({
 
 app.use('/', loginRouter);
 app.use('/', appLogicRouter);
-
-app.get('/logout', (request, response) => {
-  request.session.destroy((err) => {
-    response.redirect('/');
-  });
-});
 
 module.exports = {
   server: app,
