+++ src/routers/login-router/login.js
@@ -6,7 +6,7 @@ const express = require('express');
 const passport = require('passport');
 const Strategy = require('passport-twitter').Strategy;
 
-const createUser = require('../create-user/create-user');
+const createUser = require('../../app-logic/create-user/create-user');
 
 const loginRouter = express.Router();
 
@@ -40,6 +40,8 @@ passport.deserializeUser(function(obj, cb) {
 loginRouter.use(passport.initialize());
 loginRouter.use(passport.session());
 
+// ----------------------------------------------------------------------------
+
 loginRouter.get('/login/twitter', passport.authenticate('twitter'), (request, response) => {
   // this response should never be displayed to user if everything works correctly
   response.send({ hello: 'this is the auth route' });
@@ -60,7 +62,7 @@ loginRouter.get(
       token: request.query.oauth_token,
       token_secret: oAuthData.oauth_token_secret,
     };
-    
+
     await createUser(userData);
     request.session.destroy();
